pixel format changed from gl_rgb to gl_rgba for convenience and stuff. also unused sha1 code for later use

+++ src/main.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 
-void Initialize(GLuint &vao, const uint8_t *const pixelPtr);
+void Initialize(GLuint &vao, const uint32_t *const pixelPtr);
 GLuint CreateProgram();
 GLuint LoadAndCompileShader(const std::string &shaderName, GLenum shaderType);
 static void KeyCallback(GLFWwindow* window, int key, int scancode, int action, int mods);
